+++ BackEnd/service/src/main/java/com/basis/sge/service/servico/PerguntaServico.java
@@ -7,7 +7,6 @@ import com.basis.sge.service.servico.exception.RegraNegocioException;
 import com.basis.sge.service.servico.mapper.PerguntaMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;  //corpo da requisição, envia dados pro servidor
 
 import javax.transaction.Transactional;
@@ -32,6 +31,9 @@ public class PerguntaServico {
     }
 
     public PerguntaDTO criar(PerguntaDTO novaPergunta) {
+        if(perguntaRepositorio.existsByTituloAndObrigatoriedade(novaPergunta.getTitulo(),novaPergunta.getObrigatoriedade())){
+            throw new RegraNegocioException("Pergunta já cadastrada!");
+        }
         validaTitulo(novaPergunta.getTitulo());
         validaObrigatoriedade(novaPergunta.getObrigatoriedade());
         Pergunta pergunta = perguntaMapper.toEntity(novaPergunta);
@@ -66,7 +68,9 @@ public class PerguntaServico {
        }
     }
     public void deletar(Integer id) {
-        Pergunta pergunta = perguntaRepositorio.findById(id).orElseThrow(() -> new RegraNegocioException("Pergunta inexistente"));
+        if (!perguntaRepositorio.existsById(id)) {
+            throw new RegraNegocioException("Pergunta inexistente");
+        }
         perguntaRepositorio.deleteById(id);
     }
 }
